+++ docs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # docs
 
+## 1.3.1-canary-20250622133205
+
+### Patch Changes
+
+- Updated dependencies [5c4cd75]
+- Updated dependencies [e0b2597]
+  - @c15t/nextjs@1.3.1-canary-20250622133205
+
+## 1.3.1-canary-20250618084038
+
+### Patch Changes
+
+- Updated dependencies [95edb35]
+  - c15t@1.3.1-canary-20250618084038
+  - @c15t/backend@1.3.1-canary-20250618084038
+  - @c15t/nextjs@1.3.1-canary-20250618084038
+  - @c15t/dev-tools@1.3.1-canary-20250618084038
+  - @c15t/react@1.3.1-canary-20250618084038
+
 ## 1.3.0
 
 ### Patch Changes