+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @c15t/backend
 
+## 1.3.1-canary-20250618084038
+
+### Patch Changes
+
+- Updated dependencies [5da2f28]
+  - @c15t/translations@1.3.1-canary-20250618084038
+
 ## 1.3.0
 
 ### Minor Changes