+++ packages/dev-tools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @c15t/dev-tools
 
+## 1.3.1-canary-20250618084038
+
+### Patch Changes
+
+- Updated dependencies [95edb35]
+  - c15t@1.3.1-canary-20250618084038
+
 ## 1.3.0
 
 ### Minor Changes