+++ packages/node-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @c15t/node-sdk
 
+## 1.3.1-canary-20250618084038
+
+### Patch Changes
+
+- @c15t/backend@1.3.1-canary-20250618084038
+
 ## 1.3.0
 
 ### Minor Changes